refactor: Move security definitions processing to a separate module
